+++ controllers/question.js
@@ -1,9 +1,14 @@
-var qns = require('../json/questions.json');
+
+var serviceQns = require('../services/question.js');
 
 function getAllQuestions(req, res){
-	res.json(qns);
+	serviceQns.getAllQuestions(req, res);
+}
+function getQuestion (req, res){
+	serviceQns.getQuestion(req, res);
 }
 
 module.exports = {
-    getAllQuestions
+    getAllQuestions,
+	getQuestion
 }
